backend/requirements.txt
# Vibe-Roaster Backend Dependencies
# Python 3.11+

# ============================================================================
# Web Framework
# ============================================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ============================================================================
# Data Validation
# ============================================================================
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# ============================================================================
# Environment Management
# ============================================================================
python-dotenv==1.0.0

# ============================================================================
# AI Integration
# ============================================================================
openai==1.3.5  # Used for both Grok (via xAI API) and OpenAI GPT-4o

# ============================================================================
# Security Scanning Tools
# ============================================================================
# Note: trufflehog and semgrep should be installed separately
# pip install trufflehog semgrep bandit pip-audit

# ============================================================================
# Git Integration
# ============================================================================
gitpython==3.1.40

# ============================================================================
# Rate Limiting
# ============================================================================
slowapi==0.1.9

# ============================================================================
# HTTP Client
# ============================================================================
httpx==0.25.1

# ============================================================================
# Utilities
# ============================================================================
python-slugify==8.0.1
